fw: move select,mux*,cutoff,routing from proactive
