+++ examples/multiplexing.py
@@ -6,15 +6,15 @@ import numpy as np
 from tap import Tap
 
 from mqns.network.builder import CTRL_DELAY, NetworkBuilder
-from mqns.network.proactive import (
+from mqns.network.fw import (
     MultiplexingVector,
     MuxScheme,
     MuxSchemeBufferSpace,
     MuxSchemeStatistical,
-    ProactiveForwarder,
     QubitAllocationType,
     RoutingPathStatic,
 )
+from mqns.network.proactive import ProactiveForwarder
 from mqns.network.protocol.link_layer import LinkLayer
 from mqns.simulator import Simulator
 from mqns.utils import log, rng
