+++ examples/scalability_randomtopo_run.py
@@ -2,8 +2,9 @@ import json
 import os.path
 
 from mqns.entity.node import QNode
+from mqns.network.fw import QubitAllocationType, RoutingPathSingle
 from mqns.network.network import Request
-from mqns.network.proactive import ProactiveForwarder, ProactiveRoutingController, QubitAllocationType, RoutingPathSingle
+from mqns.network.proactive import ProactiveForwarder, ProactiveRoutingController
 from mqns.network.protocol.link_layer import LinkLayer
 from mqns.simulator import Simulator
 from mqns.utils import WallClockTimeout, json_default, log
