+++ mqns/network/builder.py
@@ -15,18 +15,17 @@ from mqns.entity.qchannel import (
 )
 from mqns.models.epr import Entanglement, MixedStateEntanglement, WernerStateEntanglement
 from mqns.models.error.input import ErrorModelInputLength
-from mqns.network.network import QuantumNetwork, TimingMode, TimingModeAsync
-from mqns.network.proactive import (
+from mqns.network.fw import (
     MuxScheme,
     MuxSchemeBufferSpace,
-    ProactiveForwarder,
-    ProactiveRoutingController,
     QubitAllocationType,
     RoutingPath,
     RoutingPathInitArgs,
     RoutingPathMulti,
     RoutingPathSingle,
 )
+from mqns.network.network import QuantumNetwork, TimingMode, TimingModeAsync
+from mqns.network.proactive import ProactiveForwarder, ProactiveRoutingController
 from mqns.network.protocol.link_layer import LinkLayer
 from mqns.network.reactive import ReactiveForwarder, ReactiveRoutingController
 from mqns.network.route import DijkstraRouteAlgorithm, RouteAlgorithm, YenRouteAlgorithm
@@ -381,17 +380,17 @@ class NetworkBuilder:
     def proactive_centralized(
         self,
         *,
-        mux: MuxScheme = MuxSchemeBufferSpace(),
+        mux: MuxScheme | None = None,
     ) -> Self:
         """
         Choose proactive forwarding with centralized control.
 
         Args:
-            mux: Multiplexing scheme.
+            mux: Multiplexing scheme, default is buffer-space.
         """
         self._assert_can_add_apps()
 
-        if isinstance(mux, MuxSchemeBufferSpace):
+        if mux is None or isinstance(mux, MuxSchemeBufferSpace):
             self.qubit_allocation = QubitAllocationType.FOLLOW_QCHANNEL
         elif isinstance(self.route, YenRouteAlgorithm):
             raise TypeError("YenRouteAlgorithm is only compatible with MuxSchemeBufferSpace")
