+++ mqns/network/fw/mux_buffer_space.py
@@ -6,14 +6,14 @@ from mqns.entity.memory import MemoryQubit, PathDirection, QubitState
 from mqns.entity.node import QNode
 from mqns.entity.qchannel import QuantumChannel
 from mqns.models.epr import Entanglement
-from mqns.network.fw import FibEntry
+from mqns.network.fw.fib import FibEntry
 from mqns.network.fw.message import PathInstructions, validate_path_instructions
-from mqns.network.proactive.mux import MuxScheme
-from mqns.network.proactive.select import MemoryEprIterator, MemoryEprTuple
+from mqns.network.fw.mux import MuxScheme
+from mqns.network.fw.select import MemoryEprIterator, MemoryEprTuple
 from mqns.utils import log, rng
 
 if TYPE_CHECKING:
-    from mqns.network.fw import Forwarder
+    from mqns.network.fw.forwarder import Forwarder
 
 
 class MuxSchemeFibBase(MuxScheme):
