+++ mqns/network/fw/mux_dynamic_epr.py
@@ -6,10 +6,10 @@ import numpy as np
 from mqns.entity.memory import MemoryQubit, QubitState
 from mqns.entity.node import QNode
 from mqns.models.epr import Entanglement
-from mqns.network.fw import Fib, FibEntry
-from mqns.network.proactive.mux_buffer_space import MuxSchemeFibBase
-from mqns.network.proactive.mux_statistical import MuxSchemeDynamicBase, has_intersect_tmp_path_ids
-from mqns.network.proactive.select import MemoryEprIterator
+from mqns.network.fw.fib import Fib, FibEntry
+from mqns.network.fw.mux_buffer_space import MuxSchemeFibBase
+from mqns.network.fw.mux_statistical import MuxSchemeDynamicBase, has_intersect_tmp_path_ids
+from mqns.network.fw.select import MemoryEprIterator
 from mqns.utils import log, rng
