+++ mqns/network/fw/mux_statistical.py
@@ -1,16 +1,19 @@
 from collections import defaultdict
 from collections.abc import Callable, Iterable, Set
-from typing import override
+from typing import TYPE_CHECKING, override
 
 from mqns.entity.memory import MemoryQubit, PathDirection, QubitState
 from mqns.entity.node import QNode
 from mqns.entity.qchannel import QuantumChannel
 from mqns.models.epr import Entanglement
-from mqns.network.fw import FibEntry, Forwarder
+from mqns.network.fw.fib import FibEntry
 from mqns.network.fw.message import PathInstructions, validate_path_instructions
-from mqns.network.proactive.mux import MuxScheme
-from mqns.network.proactive.select import MemoryEprIterator, MemoryEprTuple
+from mqns.network.fw.mux import MuxScheme
+from mqns.network.fw.select import MemoryEprIterator, MemoryEprTuple
 from mqns.utils import log, rng
+
+if TYPE_CHECKING:
+    from mqns.network.fw.forwarder import Forwarder
 
 
 def has_intersect_tmp_path_ids(epr0: Set[int] | None, epr1: Iterable[int] | None) -> bool:
