+++ mqns/network/proactive/controller.py
@@ -19,8 +19,8 @@ from typing import override
 
 from mqns.entity.cchannel import ClassicPacket
 from mqns.entity.node import Application, Controller
+from mqns.network.fw import RoutingPath
 from mqns.network.fw.message import InstallPathMsg, PathInstructions, UninstallPathMsg
-from mqns.network.proactive.routing import RoutingPath
 from mqns.utils import log
