+++ mqns/network/proactive/vora_swap.py
@@ -4,68 +4,6 @@ from typing import cast
 from mqns.entity.base_channel import default_light_speed
 from mqns.network.fw.message import SwapSequence
 
-_predefined_swap_sequences = {
-    # disable swapping (for studying isolated links)
-    "no_swap": [0, 0, 0],
-    # for 1-repeater
-    "swap_1": [1, 0, 1],
-    "swap_1_asap": [1, 0, 1],
-    # for 2-repeater
-    "swap_2_asap": [1, 0, 0, 1],
-    "swap_2_l2r": [2, 0, 1, 2],
-    "swap_2_r2l": [2, 1, 0, 2],
-    # for 3-repeater
-    "swap_3_asap": [1, 0, 0, 0, 1],
-    "swap_3_baln": [2, 0, 1, 0, 2],
-    "swap_3_l2r": [3, 0, 1, 2, 3],
-    "swap_3_r2l": [3, 2, 1, 0, 3],
-    # for 4-repeater
-    "swap_4_asap": [1, 0, 0, 0, 0, 1],
-    "swap_4_baln": [3, 0, 1, 0, 2, 3],
-    "swap_4_baln2": [3, 2, 0, 1, 0, 3],
-    "swap_4_l2r": [4, 0, 1, 2, 3, 4],
-    "swap_4_r2l": [4, 3, 2, 1, 0, 4],
-    # for 5-repeater example
-    "swap_5_asap": [1, 0, 0, 0, 0, 0, 1],
-    "swap_5_baln": [3, 0, 1, 0, 2, 0, 3],  # need to specify exact doubling  => this is used in the vora paper
-    "swap_5_baln2": [3, 0, 2, 0, 1, 0, 3],
-    "swap_5_l2r": [5, 0, 1, 2, 3, 4, 5],
-    "swap_5_r2l": [5, 4, 3, 2, 1, 0, 5],
-}
-
-
-def parse_swap_sequence(input: SwapSequence | str, route: list[str]) -> SwapSequence:
-    """
-    Parse swap sequence input.
-
-    Args:
-        input: Either an explicitly specified swap sequence,
-               or a string that identifies either a predefined swap sequence or a swapping policy.
-               The swapping policy may be one of: asap, baln, l2r, r2l.
-        route: List of nodes in the path.
-
-    Returns:
-        The swap sequence.
-
-    Raises:
-        IndexError - a predefined swap sequence is requested but not defined.
-        ValueError - specified or retrieved swap sequence does not match the route length.
-    """
-    if isinstance(input, list):
-        swap = input
-    else:
-        try:
-            swap = _predefined_swap_sequences[input]
-        except KeyError:
-            try:
-                swap = _predefined_swap_sequences[f"swap_{len(route) - 2}_{input}"]
-            except KeyError:
-                raise IndexError(f"swap sequence {input} undefined for {len(route)} nodes")
-
-    if len(swap) != len(route):
-        raise ValueError(f"swap sequence {swap} does not match route {route} with {len(route)} nodes")
-    return swap
-
 
 def compute_vora_swap_sequence(
     *,
