+++ mqns/network/reactive/forwarder.py
@@ -18,14 +18,9 @@
 from typing import override
 
 from mqns.entity.memory import MemoryQubit, PathDirection
-from mqns.network.fw import Forwarder, fw_control_cmd_handler
-from mqns.network.fw.fib import FibEntry
+from mqns.network.fw import FibEntry, Forwarder, fw_control_cmd_handler
 from mqns.network.fw.message import InstallPathMsg, UninstallPathMsg
 from mqns.network.network import TimingPhase, TimingPhaseEvent
-from mqns.network.proactive.cutoff import CutoffScheme, CutoffSchemeWaitTime
-from mqns.network.proactive.mux import MuxScheme
-from mqns.network.proactive.mux_buffer_space import MuxSchemeBufferSpace
-from mqns.network.proactive.select import SelectPurifQubit
 from mqns.network.protocol.event import ManageActiveChannels
 from mqns.network.reactive.message import LinkStateEntry, LinkStateMsg
 from mqns.utils import log
@@ -39,26 +34,6 @@ class ReactiveForwarder(Forwarder):
     It implements the forwarding phase (i.e., entanglement generation and swapping) while the centralized
     routing is done at the controller.
     """
-
-    def __init__(
-        self,
-        *,
-        ps: float = 1.0,
-        cutoff: CutoffScheme = CutoffSchemeWaitTime(),
-        mux: MuxScheme = MuxSchemeBufferSpace(),
-        select_purif_qubit: SelectPurifQubit = None,
-    ):
-        """
-        This constructor sets up a node's entanglement forwarding logic in a quantum network.
-        It configures the swapping success probability and preparing internal
-        state for managing memory, routing instructions (via FIB), synchronization,
-        and classical communication handling.
-
-        Args:
-            ps: Probability of successful entanglement swapping (default: 1.0).
-            mux: Path multiplexing scheme (default: buffer-space).
-        """
-        super().__init__(ps=ps, cutoff=cutoff, mux=mux, select_purif_qubit=select_purif_qubit)
 
     @override
     def install(self, node):
