+++ tests/proactive/test_purif.py
@@ -4,7 +4,8 @@ Test suite for ProactiveForwarder focused on purification.
 
 import pytest
 
-from mqns.network.proactive import ProactiveForwarder, RoutingPathSingle
+from mqns.network.fw import RoutingPathSingle
+from mqns.network.proactive import ProactiveForwarder
 from mqns.utils import rng
 
 from .proactive_common import (
